+++ library/src/test/java/cyclone/otusspring/library/controller/rest/BookRestControllerTest.java
@@ -2,7 +2,6 @@ package cyclone.otusspring.library.controller.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import cyclone.otusspring.library.dto.BookDto;
-import cyclone.otusspring.library.dto.BookFormDto;
 import cyclone.otusspring.library.dto.BookListElementDto;
 import cyclone.otusspring.library.dto.CommentDto;
 import cyclone.otusspring.library.mapper.AuthorMapper;
@@ -33,7 +32,8 @@ import static cyclone.otusspring.library.TestData.*;
 import static cyclone.otusspring.library.controller.rest.BookRestController.BASE_URL;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -62,12 +62,6 @@ class BookRestControllerTest {
 
     @Autowired
     BookMapper bookMapper;
-
-    @Autowired
-    AuthorMapper authorMapper;
-
-    @Autowired
-    GenreMapper genreMapper;
 
 
     @Test
@@ -159,50 +153,6 @@ class BookRestControllerTest {
                 .andExpect(status().isNoContent());
 
         verify(bookService).delete("1");
-    }
-
-    @Test
-    void getBookFormData() throws Exception {
-        when(bookService.findOne("1")).thenReturn(BOOK1);
-        when(authorService.findAll()).thenReturn(Arrays.asList(AUTHOR1, AUTHOR2));
-        when(genreService.findAll()).thenReturn(Arrays.asList(GENRE1, GENRE2));
-
-        MvcResult mvcResult =
-                mockMvc.perform(get(BASE_URL + "/formdata/1"))
-                        .andExpect(status().isOk())
-                        .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-                        .andReturn();
-
-        verify(bookService).findOne("1");
-
-        BookFormDto bookFormDto = JsonUtils.readValueFromMvcResult(objectMapper, mvcResult, BookFormDto.class);
-        assertThat(bookFormDto.getBookDto())
-                .isEqualTo(bookMapper.toBookDto(BOOK1));
-        assertThat(bookFormDto.getAllAuthors())
-                .containsExactly(authorMapper.toAuthorDto(AUTHOR1), authorMapper.toAuthorDto(AUTHOR2));
-        assertThat(bookFormDto.getAllGenres())
-                .containsExactly(genreMapper.toGenreDto(GENRE1), genreMapper.toGenreDto(GENRE2));
-    }
-
-    @Test
-    void getNewBookFormData() throws Exception {
-        when(authorService.findAll()).thenReturn(Arrays.asList(AUTHOR1, AUTHOR2));
-        when(genreService.findAll()).thenReturn(Arrays.asList(GENRE1, GENRE2));
-
-        MvcResult mvcResult =
-                mockMvc.perform(get(BASE_URL + "/formdata/new"))
-                        .andExpect(status().isOk())
-                        .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-                        .andReturn();
-
-        verify(bookService, never()).findOne(any());
-
-        BookFormDto bookFormDto = JsonUtils.readValueFromMvcResult(objectMapper, mvcResult, BookFormDto.class);
-        assertThat(bookFormDto.getBookDto()).isNull();
-        assertThat(bookFormDto.getAllAuthors())
-                .containsExactly(authorMapper.toAuthorDto(AUTHOR1), authorMapper.toAuthorDto(AUTHOR2));
-        assertThat(bookFormDto.getAllGenres())
-                .containsExactly(genreMapper.toGenreDto(GENRE1), genreMapper.toGenreDto(GENRE2));
     }
 
     @Test
